// web/src/api/generated/model/trashedFile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { TrashedFileEncryptionMode } from "./trashedFileEncryptionMode.ts";

export interface TrashedFile {
  delete_after: string;
  encrypted_name?: string;
  encryption_mode: TrashedFileEncryptionMode;
  id: string;
  library_id: string;
  library_name: string;
  name?: string;
  name_token?: string;
  parent_id: string;
  revision: number;
  trashed_at: string;
}